requirements.txt
streamlit>=1.28.0
mysql-connector-python>=8.2.0
bcrypt>=4.0.1
googletrans==4.0.0rc1
openai-whisper>=20231117
speechrecognition>=3.10.0
pydub>=0.25.1
pillow>=10.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
streamlit-geolocation>=0.1.0
streamlit-authenticator>=0.2.3
streamlit-camera-input-live>=0.2.0
indic-transliteration>=2.3.0
deep-translator>=1.11.4
streamlit-option-menu>=0.3.6
streamlit-extras>=0.3.0
